+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed `table` view to use `<table>` tag for root element instead of `<div>`
+- Added `debounce` property to `input` view config.
 
 ## 1.0.0-beta.31 (23-04-2020)