+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fixed `table` view to use `<table>` tag for root element instead of `<div>`
 - Added `debounce` property to `input` view config.
+- Exposed lib API as `Widget#lib`
+- Extended `safeFilterRx()` to add pattern value to produced regexp as `rawSource`
 
 ## 1.0.0-beta.31 (23-04-2020)