+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Extended protocol for `page.options.encodeParams()`, handler may return an array of URI component pairs or an object beside a string; keys and values will be encoded by host
 - Changed references to default page as `defaultPageId` instead of hardcoded `'default'` value
 - Added `Widget.options.reportPageId` option
+- Improved `Widget.setPageHash()` method to detect changes based on decoded page id, ref and params and re-apply new values if needed
 
 ## 1.0.0-beta.31 (23-04-2020)