+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - Changed references to default page as `defaultPageId` instead of hardcoded `'default'` value
 - Added `Widget.options.reportPageId` option
 - Improved `Widget.setPageHash()` method to detect changes based on decoded page id, ref and params and re-apply new values if needed
-- Fixed `table` to not override `data` when col config in `cols` object have no `content` property
+- Fixed `table` to not override `data` when column config in `cols` object have no `content` property
+- Fixed `table` to apply `when` on column config to entire column
 
 ## 1.0.0-beta.31 (23-04-2020)